Adjust min and max values in client
Make sure we serialize tppFloat entirely instead of going with a property:value approach. The version system can take care of any eventual differences. Delete PropertyHeader and the enum
Fix bug in reception of packets on the server side not taking into account the Message Header
Put min and max in the Slider widget
Disable context menu in body