Rework how variables get updated. The server now stores the hash the client has provided, and sends it back on update
This means we don't need to send the entire path every time we update a variable, which is a big saving. It also means we don't hash it on the client
This saving will be useful when we add bidirectional variables, i.e. variables that the client is able to modify
One thing that this helps with as well is to start separating the variable description from the variable itself, which is a tangled concept at the moment
The view of the server and the client need to be radically different, but they influence each other in ways we don't want